v1.4.1 change external snapshot

+++ operation.py
@@ -1089,8 +1089,9 @@ def deleteExternalSnapshot(params):
                         op = Operation('qemu-img rebase -b %s %s' % (bf_bf_path, disk_config['current']), {})
                         op.execute()
                 # delete backing file
-                op = Operation('rm -f %s' % params.backing_file, {})
-                op.execute()
+                if DiskImageHelper.get_backing_file(disk_config['current']) != params.backing_file:
+                    op = Operation('rm -f %s' % params.backing_file, {})
+                    op.execute()
                 # for df in snapshots_to_delete:
                 #     op = Operation('rm -f %s' % df, {})
                 #     op.execute()
@@ -1105,8 +1106,9 @@ def deleteExternalSnapshot(params):
                                    (params.domain, disk_config['current']), {})
                     op.execute()
                 # delete backing file
-                op = Operation('rm -f %s' % params.backing_file, {})
-                op.execute()
+                if DiskImageHelper.get_backing_file(disk_config['current']) != params.backing_file:
+                    op = Operation('rm -f %s' % params.backing_file, {})
+                    op.execute()
             # modify json file, make os_event_handler to modify data on api server .
             with open(disk_config['dir'] + '/config.json', "r") as f:
                 config = load(f)
